requirements.txt
alembic==1.13.3
asgiref==3.8.1
black==24.10.0

Django==5.1.1
django-debug-toolbar==4.4.6

Jinja2==3.1.4
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0

pillow==11.0.0
PyJWT==2.3.0
pyparsing==2.4.7
pytz==2024.2
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.5.1
django-durationwidget==1.0.5